+++ projects/js-packages/publicize-components/src/components/schedule-button/stories/index.stories.jsx
@@ -1,0 +1,14 @@
+import ScheduleButton from '../index.tsx';
+
+export default {
+	title: 'JS Packages/Publicize Components/ScheduleButton',
+	component: ScheduleButton,
+};
+
+const Template = args => <ScheduleButton { ...args } />;
+
+export const _default = Template.bind( {} );
+_default.args = {
+	scheduleTimestamp: Date.now() / 1000,
+	isBusy: false,
+};