+++ projects/js-packages/publicize-components/src/social-store/hydrate-stores.ts
@@ -60,4 +60,15 @@ export async function hydrateStores() {
 
 		await finishResolution( 'getEntityRecords', [ 'wpcom/v2', 'publicize/services' ] );
 	}
+
+	if ( ! wpcomEntities.some( ( { name } ) => name === 'publicize/scheduled-actions' ) ) {
+		await addEntities( [
+			{
+				kind: 'wpcom/v2',
+				name: 'publicize/scheduled-actions',
+				baseURL: '/wpcom/v2/publicize/scheduled-actions',
+				label: __( 'Publicize scheduled actions', 'jetpack-publicize-components' ),
+			},
+		] );
+	}
 }