+++ projects/plugins/jetpack/extensions/plugins/seo/index.js
@@ -16,6 +16,7 @@ import { PluginPrePublishPanel } from '@wordpress/edit-post';
 import { store as editorStore } from '@wordpress/editor';
 import { createPortal } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
+import clsx from 'clsx';
 /**
  * Internal dependencies
  */
@@ -129,18 +130,26 @@ const Seo = () => {
 							<SeoAssistantSidebarEntrypoint disabled={ false } placement="jetpack-sidebar" />
 						</PanelRow>
 					) }
-					{ isSeoEnhancerEnabled && isViewable && (
-						<PanelRow>
-							<SeoEnhancer />
-						</PanelRow>
-					) }
-					<PanelRow>
+					{ isSeoEnhancerEnabled && isViewable && <SeoEnhancer /> }
+					<PanelRow
+						className={ clsx( {
+							'jetpack-seo-sidebar__feature-section': isSeoEnhancerEnabled,
+						} ) }
+					>
 						<SeoTitlePanel />
 					</PanelRow>
-					<PanelRow>
+					<PanelRow
+						className={ clsx( {
+							'jetpack-seo-sidebar__feature-section': isSeoEnhancerEnabled,
+						} ) }
+					>
 						<SeoDescriptionPanel />
 					</PanelRow>
-					<PanelRow>
+					<PanelRow
+						className={ clsx( {
+							'jetpack-seo-sidebar__feature-section': isSeoEnhancerEnabled,
+						} ) }
+					>
 						<SeoNoindexPanel />
 					</PanelRow>
 				</PanelBody>
@@ -148,11 +157,7 @@ const Seo = () => {
 
 			<PluginPrePublishPanel { ...jetpackSeoPrePublishPanelProps }>
 				<>
-					{ isSeoEnhancerEnabled && isViewable && (
-						<PanelRow>
-							<SeoEnhancer />
-						</PanelRow>
-					) }
+					{ isSeoEnhancerEnabled && isViewable && <SeoEnhancer /> }
 					<PanelRow>
 						<SeoTitlePanel />
 					</PanelRow>